re #9 - playable and class diagram (version 1)
This game would be a mix of asteroids and a 2D platformer. I mde sure that the basic classes, a character that can shoot and the asteroids are present. You can aim and shoot with the mouse at the asteroid which will split into more pieces until they are too small and will disappear.