Added: - IDamagable WIP.
Edited:
Block.cs
Does work with events now and setting the damage of the block.
BlockGenerator.cs
Generating his own block at the beginning and setting the block his event.
InputManager
only provide the Game with input now and it's not dependent on the block class
Singleton.cs
UML-Geoffrey-Hendrikx.png