Prototype files to test out the new UIElements system.
NOTE:
These files are not mine. I did create my own UIElements Editor, but ultimately scrapped it as I found an article were Unity said that the system was still unstable, and some features in the documentation might still not even work (as I encountered when trying to create an EnumField).